Host Report documentation and linting.
